requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic>=2.10.6,<3
python-dotenv==1.0.1

# Pipecat - use latest version without pinning extras versions
pipecat-ai[daily,google,silero]==0.0.87
onnxruntime==1.20.1
loguru==0.7.3

# Let pip resolve compatible versions for these
aiohttp
google-genai

# Deepgram for language detection and STT
deepgram-sdk>=3.0.0
